Move roboflow API utils to separate module
